requirements.txt
PyYAML
dotenv
